订单多图上传

+++ cofactory-1.1/Utils/SDPhotoBrowser/SDPhotoBrowser.h
@@ -27,7 +27,7 @@
 @interface SDPhotoBrowser : UIView <UIScrollViewDelegate>
 
 @property (nonatomic, weak) UIView *sourceImagesContainerView;
-@property (nonatomic, assign) int currentImageIndex;
+@property (nonatomic, assign) NSUInteger currentImageIndex;
 @property (nonatomic, assign) NSInteger imageCount;
 
 @property (nonatomic, weak) id<SDPhotoBrowserDelegate> delegate;
